+++ Lights/SpotLight.h
@@ -12,12 +12,12 @@ public:
 	float cutoffAngle;
 	float outerCutoffAngle;
 
-	SpotLight(uint8_t &lightNum, std::vector<PointLight*> &container, uint32_t &objectNum)
+	SpotLight(uint16_t &lightNum, std::vector<PointLight*> &container, uint32_t &objectNum)
 		: PointLight(lightNum, container, objectNum),
 		direction(glm::vec3(0.0, -1.0, 0.0)), 
 		cutoffAngle(40.0f), 
 		outerCutoffAngle(45.0f){}
-	SpotLight(uint8_t &lightNum, std::vector<PointLight*>& container, uint32_t &objectNum, glm::vec3 position, glm::vec3 direction, float cutoffAngle, float outerCutoffAngle, float constant, float linear, float quadratic, glm::vec3 ambient, glm::vec3 diffuse, glm::vec3 specular)
+	SpotLight(uint16_t &lightNum, std::vector<PointLight*>& container, uint32_t &objectNum, glm::vec3 position, glm::vec3 direction, float cutoffAngle, float outerCutoffAngle, float constant, float linear, float quadratic, glm::vec3 ambient, glm::vec3 diffuse, glm::vec3 specular)
 		:PointLight(lightNum, container, objectNum, position, constant, linear, quadratic, ambient, diffuse, specular),
 		direction(direction), 
 		cutoffAngle(cutoffAngle),
